addressing moving assertion outsitde of custom commands and various updates
-renamed custom commands files
-refactored so that returned a result and moved assertion to calling test
-added @role=button to various page objects
-replaced === in various comparisons
