+++ pageobjects/clinicPatientListPage.js
@@ -10,35 +10,35 @@ module.exports = {
 
         chartDateRangeModal: '#ChartDateRangePicker',
         patientDetails: {
-          selector: '//*[@id="peopleTable-header-fullName"]//span',
+          selector: '//*[@id="peopleTable-header-fullName"]//span[@role="button"]',
           locateStrategy: 'xpath',
         },
         lastUpload: {
-          selector: '//*[@id="peopleTable-header-cgm-lastUploadDate"]//span',
+          selector: '//*[@id="peopleTable-header-cgm-lastUploadDate"]//span[@role="button"]',
           locateStrategy: 'xpath',
         },
         patientTags: {
-          selector: '//*[@id="peopleTable-header-tags"]//span',
+          selector: '//*[@id="peopleTable-header-tags"]//span[@role="button"]',
           locateStrategy: 'xpath',
         },
         GMI: {
-          selector: '//*[@id="peopleTable-header-cgm-glucoseManagementIndicator"]//span',
+          selector: '//*[@id="peopleTable-header-cgm-glucoseManagementIndicator"]//span[@role="button"]',
           locateStrategy: 'xpath',
         },
         timeInRange: {
-          selector: '//*[@id="peopleTable-header-bgRangeSummary"]//span',
+          selector: '//*[@id="peopleTable-header-bgRangeSummary"]//span[@role="button"]',
           locateStrategy: 'xpath',
         },
         avgGlucose: {
-          selector: '//*[@id="peopleTable-header-bgm-averageGlucoseMmol"]//span',
+          selector: '//*[@id="peopleTable-header-bgm-averageGlucoseMmol"]//span[@role="button"]',
           locateStrategy: 'xpath',
         },
         lows: {
-          selector: '//*[@id="peopleTable-header-bgm-timeInVeryLowRecords"]//span',
+          selector: '//*[@id="peopleTable-header-bgm-timeInVeryLowRecords"]//span[@role="button"]',
           locateStrategy: 'xpath',
         },
         highs: {
-          selector: '//*[@id="peopleTable-header-bgm-timeInVeryHighRecords"]//span',
+          selector: '//*[@id="peopleTable-header-bgm-timeInVeryHighRecords"]//span[@role="button"]',
           locateStrategy: 'xpath',
         },
         showAll: {
@@ -46,7 +46,7 @@ module.exports = {
           locateStrategy: 'xpath',
         },
         lastUploadDesc: {
-          selector: '//*[@id="peopleTable-header-cgm-lastUploadDate"]//*[contains(@class," MuiTableSortLabel-iconDirectionDesc")]',
+          selector: '//*[@id="peopleTable-header-cgm-lastUploadDate"]//span[@role="Button"]',
           locateStrategy: 'xpath',
         },
         workspaceSettings: {
